Float times on ringer wire (#7)
* fix: parse float on ringer wire duration

* feat: release 1.1.2

+++ src/utils/ring.ts
@@ -17,7 +17,7 @@ export const ring = async (
     }
   }
 
-  const steps = ringerWire.split(',').map(x => parseInt(x))
+  const steps = ringerWire.split(',').map(x => parseFloat(x))
 
   let action: High | Low = 1
   let count = 0
